+++ CHANGELOG.md
@@ -15,6 +15,12 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Bug fixes
+
+- The `build` script in watch mode no longer hangs without building.
+
 ## 1.0.0-alpha.3 (February 11, 2019)
 
 ### Bug fixes